wave2

+++ Assets/ScriptableObjects/Monster/Scripts/MonStatData.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "New Stat Data", menuName = "ScriptableObjects/Monster/Stat Data")]
 public class MonStatData : ScriptableObject
 {
-    public int maxHp;
+    public float maxHp;
     public int defense;
     public float moveSpeed;
 }
